+++ tests/test_decorator.py
@@ -9,6 +9,9 @@ from test_flake8_trio import _default_option_manager
 
 from flake8_trio import Error_codes, Plugin, Statement, fnmatch_qualified_name
 
+if sys.version_info[:2] < (3, 9):
+    pytest.skip(allow_module_level=True)
+
 
 def dec_list(*decorators: str) -> ast.Module:
     source = ""